==> Mouse_SI_GW_SC/07_integration_IMM_DN.Rmd <==
---
title: "`r params$project_title`"
editor_options:
  chunk_output_type: console
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  bookdown::html_document2:
    code_folding: hide
    toc: yes
    toc_float:
      collapsed: no
  pdf_document:
    toc: yes
params:
  dataset: 1
---

```{r Setup, echo = FALSE}
knitr::opts_chunk$set(fig.width  = 10,
                      results    = "hold",
                      message    = FALSE,
                      warning    = FALSE)


```

```{r setup, include=FALSE}



suppressPackageStartupMessages({
    library(Seurat)
    library(cowplot)
    library(ggplot2)
})


```



# Integrate data

```{r}


library(future)
plan("multiprocess", workers = 20)
options(future.globals.maxSize = 120000 * 1024^2)

print("IMM_DN")


if(!file.exists("../analysis/filtered_IMM_DN_int.rds")){
  if(!file.exists("../analysis/IMM_DN.anchors.rds")){
    
    IMM_DN.all <- readRDS("../analysis/filtered_IMM_DN_dr.rds")
    IMM_DN.all <- AddMetaData(IMM_DN.all, gsub("IMM","", gsub("DN","",IMM_DN.all$orig.ident)),"sample")
    
    IMM_DN.list <- SplitObject(IMM_DN.all, split.by = "sample")
    
    x = IMM_DN.list
    for (i in 1:length(x)) {
        x[[i]] <- NormalizeData(x[[i]], verbose = FALSE)
        x[[i]] <- FindVariableFeatures(x[[i]], selection.method = "vst",
            nfeatures = 2000, verbose = FALSE)
    }
    IMM_DN.list = x
    
    IMM_DN.anchors <- FindIntegrationAnchors(object.list = IMM_DN.list, dims = 1:30, 
        reduction = "cca")
    
    saveRDS(IMM_DN.anchors,"../analysis/IMM_DN.anchors.rds")
    
    
    rm(IMM_DN.list)
    gc()
  }else{
    
    IMM_DN.all <- readRDS("../analysis/filtered_IMM_DN_clus_dr.rds")
    IMM_DN.anchors = readRDS("../analysis/IMM_DN.anchors.rds")
  
  }
  
  IMM_DN.int <- IntegrateData(anchorset = IMM_DN.anchors, dims = 1:30, new.assay.name = "CCA")

  saveRDS(IMM_DN.int, "../analysis/filtered_IMM_DN_int.rds")
  
  rm(IMM_DN.anchors)
  gc()
  
  IMM_DN.int <- ScaleData(IMM_DN.int)
  #IMM_DN.int <- ScaleData(IMM_DN.int, assay = "RNA")
  IMM_DN.int <- RunPCA(IMM_DN.int, npcs = 30, verbose = FALSE)
  IMM_DN.int <- RunUMAP(IMM_DN.int, dims = 1:30)
  #IMM_DN.int <- RunTSNE(IMM_DN.int, dims = 1:30)
  
  saveRDS(IMM_DN.int, "../analysis/filtered_IMM_DN_int.rds")
}else{
  IMM_DN.int = readRDS("../analysis/filtered_IMM_DN_int.rds")
  IMM_DN.all <- readRDS("../analysis/filtered_IMM_DN_dr.rds")
}


plot_grid(ncol = 2,
  DimPlot(IMM_DN.all, reduction = "pca", group.by = "orig.ident", 
          pt.size = 0.01, shuffle = TRUE) +
    NoAxes() + ggtitle("PCA raw_data"),
  DimPlot(IMM_DN.all, reduction = "umap", group.by = "orig.ident", 
          pt.size = 0.01, shuffle = TRUE) +
    NoAxes() + ggtitle("UMAP raw_data"),
  
  DimPlot(IMM_DN.int, reduction = "pca", group.by = "orig.ident",
          shuffle = TRUE) +
    NoAxes() + ggtitle("PCA integrated"),
  DimPlot(IMM_DN.int, reduction = "umap", group.by = "orig.ident",
          shuffle = TRUE) +
    NoAxes() + ggtitle("UMAP integrated")
)



```

# Plot QC in UMAP

```{r, fig.height = 10}

FeaturePlot(IMM_DN.int, reduction = "umap", features = c("percent_mito","percent_ribo","S.Score", "G2M.Score", "nFeature_RNA"), 
    order = T, slot = "data", combine = T)

```



# Plot genes of interest

```{r}

DefaultAssay(IMM_DN.int) = "RNA"

FeaturePlot(IMM_DN.int, reduction = "umap", features = c("Mki67","Pcna"), 
    order = T, slot = "data", combine = T)

FeaturePlot(IMM_DN.int, reduction = "umap", features = c("Ptprc","Epcam"), 
    order = T, slot = "data", combine = T)


```

```{r}

sessionInfo()

```
